rename encryption resource
