+++ service/controller/resource/tccpencryption/delete.go
@@ -1,4 +1,4 @@
-package encryption
+package tccpencryption
 
 import (
 	"context"
